pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = "Password wallet manager"
authors = []

[tool.poetry.dependencies]
python = "^3.9"
asyncpg = "^0.25.0"
uvicorn = "^0.17.5"
uvloop = "^0.16.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.1"}
aerich = "^0.6.2"
fastapi-jwt-auth = "^0.5.0"
bcrypt = "^3.2.0"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
ipython = "^7.29.0"
mypy = "^0.910"
python-semantic-release = "^7.25.0"
httpx = "^0.22.0"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.18.2"
pytest-cov = "^3.0.0"
isort = "^5.10.1"


[tool.aerich]
tortoise_orm = "wallet.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./wallet"

[tool.pytest.ini_options]
addopts = "--cov=./wallet"
testpaths = [
   "./tests"
]
asyncio_mode = "auto"

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true

[tool.semantic_release]
version_variable = [
    'wallet/_version.py:__version__'
]
version_toml = 'pyproject.toml:tool.poetry.version'
upload_to_repository=false
major_on_zero=true
hvcs = "gitlab"
commit_message = "Bump version: {version}"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"